+++ R/predict.R
@@ -164,7 +164,7 @@
 #' unique(d$year)
 #' m <- sdmTMB(
 #'   data = d, formula = density ~ 1,
-#'   ar1_fields = TRUE, # using an AR1 to have something to forecast with
+#'   fields = "AR1", # using an AR1 to have something to forecast with
 #'   extra_time = 2019L,
 #'   include_spatial = FALSE,
 #'   time = "year", spde = pcod_spde, family = tweedie(link = "log")