+++ R/priors.R
@@ -154,7 +154,7 @@ halfnormal <- function(location = 0, scale = 1) {
 #' # (except on the annual means; see the `NA`s)
 #' m <- sdmTMB(density ~ 0 + depth_scaled + depth_scaled2 + as.factor(year),
 #'   data = d, time = "year", spde = pcod_spde, family = tweedie(link = "log"),
-#'   share_range = FALSE, ar1_fields = TRUE,
+#'   share_range = FALSE, fields = "AR1",
 #'   priors = sdmTMBpriors(
 #'     b = normal(c(0, 0, NA, NA, NA), c(2, 2, NA, NA, NA)),
 #'     phi = halfnormal(0, 10),