+++ R/sim.R
@@ -58,7 +58,7 @@
 #' m <- sdmTMB(
 #'   data = s, formula = observed ~ x1,
 #'   time = "time", spde = mesh,
-#'   ar1_fields = TRUE, include_spatial = FALSE
+#'   fields = "AR1", include_spatial = FALSE
 #' )
 #' tidy(m, conf.int = TRUE)
 #' tidy(m, "ran_pars", conf.int = TRUE)