+++ tests/testthat/test-priors.R
@@ -27,13 +27,13 @@ test_that("Prior fitting works", {
   # no priors
   m <- sdmTMB(density ~ 0 + depth_scaled + depth_scaled2 + as.factor(year),
     data = d, time = "year", spde = pcod_spde, family = tweedie(link = "log"),
-    ar1_fields = TRUE,
+    fields = "AR1",
     share_range = FALSE)
 
   # all the priors
   mp <- sdmTMB(density ~ 0 + depth_scaled + depth_scaled2 + as.factor(year),
     data = d, time = "year", spde = pcod_spde, family = tweedie(link = "log"),
-    share_range = FALSE, ar1_fields = TRUE,
+    share_range = FALSE, fields = "AR1",
     priors = sdmTMBpriors(
       b = normal(c(0, 0, NA, NA, NA), c(2, 2, NA, NA, NA)),
       phi = halfnormal(0, 10),